+++ addOns/scripts/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Update minimum ZAP version to 2.9.0.
 - Tweak help content.
+- Show script engine when editing the script.
 
 ## [26] - 2020-01-17
 ### Added